Data augmentation client visualization (work in progress) + bash script to generate jdoctor conditions

+++ data-augmentation/migrate.sh
@@ -15,29 +15,38 @@
 current_dir="$(realpath "$(dirname "$BASH_SOURCE")")"
 
 # Get global variables
-source "${current_dir}/utils/global_variables.sh"
+source "${current_dir}/generator/utils/global_variables.sh"
 
 # Define local variables
 data_aug_resources_dir="${current_dir}/src/main/resources"
-projects_source_dir="${data_aug_resources_dir}/projects_source"
+projects_source_dir="${data_aug_resources_dir}/projects-source"
 
 python3 "${ROOT_DIR}/scripts/generate_input_projects.py"
 
 # Read the CSV file line by line and split into project_name, github_link, and commit fields
 while IFS=, read -r project_name github_link commit; do
-  echo "Cloning project: ${project_name}"
+  # Clean commit string from undesired white-spaces/line-breaks introduced with the CSV parsing
+  commit="${commit//[$'\t\r\n ']/}"
+  # Create path if it does not exists
+  if [ ! -d "$projects_source_dir" ]; then
+      mkdir -p "$projects_source_dir";
+  fi
   # Move to the projects_source directory where to clone the current project
   cd "$projects_source_dir"
-  # Clone the project
-  git clone "$github_link" "$project_name"
-  # Check if a specific commit must be set
-  if [ "$commit" != "last" ]; then
-    echo "Move head to commit: ${commit}"
-    # Move within the project
-    cd "${projects_source_dir}/${project_name}"
-    git reset --hard "$commit"
+  # Check if the project has already been cloned
+  if [ ! -d "${projects_source_dir}/${project_name}" ]; then
+    echo "Cloning project: ${project_name}"
+    # Clone the project
+    git clone "$github_link" "$project_name"
+    # Check if a specific commit must be set
+      if [ "$commit" != "last" ]; then
+        echo "Move head to commit: ${commit}"
+        # Move within the project
+        cd "${projects_source_dir}/${project_name}"
+        git reset --hard "$commit"
+      fi
   fi
-done < "$INPUT_PROJECTS_CSV_FILE}"
+done < "$INPUT_PROJECTS_CSV_FILE"
 # Move to the root directory
 cd "$current_dir"
 # Generate collections and documents for each project
